% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{acc_from_fasta}
\alias{acc_from_fasta}
\title{Genbank accessions from fasta}
\usage{
acc_from_fasta(x)
}
\arguments{
\item{x}{A filepath or vector of filepaths to fasta files}
}
\description{
Genbank accessions from fasta
}
