% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbdemux}
\alias{bbdemux}
\title{Demultiplex fusion primers using BBmap Seal}
\usage{
bbdemux(
  install = NULL,
  fwd,
  rev = NULL,
  Fbarcodes = NULL,
  Rbarcodes = NULL,
  restrictleft = NULL,
  out.dir = "demux",
  kmer = NULL,
  hdist = 0,
  degenerate = TRUE,
  overwrite = TRUE,
  threads = NULL,
  mem = NULL,
  interleaved = FALSE
)
}
\arguments{
\item{install}{(Required) Install location for bbmap}

\item{fwd}{(Required) Vector of locations of forward reads}

\item{rev}{(Optional) Vector of locations of reverse reads}

\item{Fbarcodes}{(Required) Barcodes used in forward reads}

\item{Rbarcodes}{(Optional) Barcodes used in reverse reads}

\item{restrictleft}{(Optional) Defaults to the size of the largest primer.
Restricts the kmer search for primer sequences to just the left side of the molecule.}

\item{out.dir}{(Optional) Default "demux"
The path to write the output reads.}

\item{kmer}{(Optional) default the size of the smallest primer will be used.
The kmer size to use for primer searching.}

\item{hdist}{(Optional) Default = 0. The hamming distance (number of substitution errors) allowed for mismatch to the query primer.}

\item{degenerate}{(Optional) Default TRUE.
Option to search for all possible primer combinations for degenerate primers}

\item{overwrite}{(Optional) Default TRUE
Option to overwrite existing output files.}

\item{threads}{(Optional) Default autodetect
Number of CPU threads to use
WARNING: Thread detection can fail on cluster computing currently}

\item{mem}{(Optional) Default autodetect
GB of memory to use
WARNING: mem detection can fail on cluster computing currently}

\item{interleaved}{(Optional) Default FALSE
Option to input interleaved reads}
}
\value{

}
\description{
Demultiplex fusion primers using BBmap Seal
}
\examples{
#' \dontrun{
path <- "run_test/"
demuxpath <- file.path(path, "demux") # Filtered forward files go into the path/filtered/ subdirectory

fastqFs <- sort(list.files(path, pattern="R1_001.*", full.names = TRUE))
fastqRs <- sort(list.files(path, pattern="R2_001.*", full.names = TRUE))

bbdemux(install="bin/bbmap", fwd=fastqFs, rev=fastqRs,Fbarcodes = c("GAGGDACW","TGTGGDAC","AGAAGGDAC"),
              Rbarcodes = c("ACGTRATW","TCCGTRAT","CTGCGTRA"),
              degenerate=TRUE, out.dir=demuxpath, threads=1, mem=4,
              hdist=0, overwrite=TRUE)
              }

}
