% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbinstall}
\alias{bbinstall}
\title{Install BBtools}
\usage{
bbinstall(url, dest.dir = "bin")
}
\arguments{
\item{url}{(Optional) Default will search for the latest version
URL to retrieve bbmap from.}

\item{dest.dir}{(Optional)  Default "bin"
Directory to install bbmap within.}
}
\description{
Install BBtools
}
