% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbmap_install}
\alias{bbmap_install}
\title{BBMap Install}
\usage{
bbmap_install(url, destdir = "bin")
}
\arguments{
\item{url}{(Optional) Default = "https://downloads.sourceforge.net/project/bbmap/BBMap_38.57.tar.gz"
URL to retrieve bbmap from.}

\item{destdir}{(Optional)  Default "bin"
Directory to install bbmap within.}
}
\description{
BBMap Install
}
