% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbsplit}
\alias{bbsplit}
\title{Split interleaved reads}
\usage{
bbsplit(install = NULL, files, overwrite = FALSE)
}
\arguments{
\item{install}{(Required) Install location for bbmap}

\item{files}{(Required) Vector of locations of interleaved read files to split}

\item{overwrite}{(Optional) Default TRUE
Option to overwrite existing output files.}
}
\value{

}
\description{
Split interleaved reads
}
