% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbtools_demux}
\alias{bbtools_demux}
\title{Demultiplex fusion primers using BBmap Seal}
\usage{
bbtools_demux(install = NULL, fwds, revs = NULL, Fbarcodes = NULL,
  Rbarcodes, restrictleft = NULL, outpath = "trimmed", kmer = NULL,
  hdist = 0, copyundefined = TRUE, overwrite = TRUE,
  interleaved = FALSE)
}
\arguments{
\item{interleaved}{}
}
\description{
Demultiplex fusion primers using BBmap Seal
}
