% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbtools_split}
\alias{bbtools_split}
\title{Split interleaved reads}
\usage{
bbtools_split(install = NULL, files, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
Split interleaved reads
}
