% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbtools_trim}
\alias{bbtools_trim}
\title{Trim primers using BBDuk}
\usage{
bbtools_trim(install = NULL, fwd, rev = NULL, primers,
  restrictleft = NULL, outpath = "bbduk", ktrim = "l",
  ordered = TRUE, kmer = NULL, mink = FALSE, tpe = TRUE,
  hdist = 0, copyundefined = TRUE, overwrite = TRUE,
  quality = TRUE, maxlength = NULL)
}
\arguments{
\item{maxlength}{}
}
\description{
Trim primers using BBDuk
}
