% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{blast}
\alias{blast}
\title{Run BLAST search}
\usage{
blast(query, db, type = "blastn", evalue = 1e-06, args = NULL,
  quiet = FALSE)
}
\arguments{
\item{query}{Query sequence. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.}

\item{db}{Reference sequences to conduct search against. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.
If DNAbin, DNAStringSet or character string is provided, a temporary fasta file is used to construct BLAST database}

\item{type}{type of search to conduct, default 'blastn'}

\item{evalue}{Minimum evalue from search}

\item{args}{Extra arguments passed to BLAST}

\item{quiet}{Whether progress should be printed to console}
}
\description{
Run BLAST search
}
