% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{blast_install}
\alias{blast_install}
\title{Install BLAST}
\usage{
blast_install(url, destdir = "bin")
}
\arguments{
\item{url}{A URL for the specific BLAST release to download. If missing, automatically download latest release.}

\item{destdir}{Destination to install BLAST to}
}
\description{
Install BLAST
}
