% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{blast_params}
\alias{blast_params}
\title{Show BLAST parameters}
\usage{
blast_params(type = "blastn")
}
\arguments{
\item{type}{(Required) Type of blast to display help page for}
}
\description{
Show BLAST parameters
}
