% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_taxa (Alexander’s MacBook Pro's
%   conflicted copy 2019-06-14).R, R/fetch_taxa.R
\name{boldSearch}
\alias{boldSearch}
\title{boldSearch}
\usage{
boldSearch(x, marker = "COI-5P", quiet = FALSE, output = "h",
  file = NULL, compress = FALSE, dir = NULL)

boldSearch(x, marker = "COI-5P", quiet = FALSE, output = "h",
  file = NULL, compress = FALSE, dir = NULL)
}
\arguments{
\item{dir}{}

\item{dir}{}
}
\description{
boldSearch

boldSearch
}
