% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{clean_seqs}
\alias{clean_seqs}
\title{Clean Sequences with PHMM}
\usage{
clean_seqs(x, model, minscore = 100, minamplen = 50, maxamplen = 500,
  shave = FALSE, maxNs = 0.02, cores = 1, quiet = FALSE)
}
\arguments{
\item{quiet}{}
}
\description{
Clean Sequences with PHMM
}
