% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{codon_entropy}
\alias{codon_entropy}
\title{Codon entropy}
\usage{
codon_entropy(
  x,
  genetic_code = NULL,
  forward = TRUE,
  reverse = FALSE,
  codon_filter = TRUE,
  method = "ML"
)
}
\arguments{
\item{genetic_code}{A genetic code for the Amino acid translation. set to 'SGC4' for Invertebrate mitochondrial or see all known codes at Biostrings::GENETIC_CODE_TABLE}

\item{codon_filter}{}
}
\value{

}
\description{
Codon entropy
}
