% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{codon_filter}
\alias{codon_filter}
\title{Filter sequences containing stop codons}
\usage{
codon_filter(
  x,
  genetic_code = NULL,
  forward = TRUE,
  reverse = FALSE,
  remove_ambiguities = TRUE
)
}
\arguments{
\item{x}{Sequences in DNAStringset or DNAbin format}

\item{genetic_code}{A genetic code for the Amino acid translation. set to 'SGC4' for Invertebrate mitochondrial or see all known codes at Biostrings::GENETIC_CODE_TABLE}

\item{forward}{Whether the forward complement should be used}

\item{reverse}{Whether the reverse complement should be used}

\item{remove_ambiguities}{Whether sequences containing ambiguous bases (non ATGC) should be removed}
}
\value{

}
\description{
Filter sequences containing stop codons
}
