% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAbin_utilities.R
\name{concat_DNAbin}
\alias{concat_DNAbin}
\title{Concatenate DNAbin objects while preserving attributes.
This function joins two or more \code{DNAbin} objects, retaining any
attributes whose lengths match those of the input objects (e.g. "species",
"lineage" and/or "taxID" attributes).}
\usage{
concat_DNAbin(...)
}
\arguments{
\item{...}{\code{DNAbin} objects, or list of DNAbins to be concatenated.}
}
\value{

}
\description{
Concatenate DNAbin objects while preserving attributes.
This function joins two or more \code{DNAbin} objects, retaining any
attributes whose lengths match those of the input objects (e.g. "species",
"lineage" and/or "taxID" attributes).
}
