% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_samplesheet}
\alias{create_samplesheet}
\title{Create Samplesheet}
\usage{
create_samplesheet(SampleSheet, runParameters, format = "miseq")
}
\arguments{
\item{Rtwintag}{}
}
\value{

}
\description{
Create Samplesheet
}
