% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{DNAbin2DNAstringset}
\alias{DNAbin2DNAstringset}
\title{Convert DNABin to DNAStringSet}
\usage{
DNAbin2DNAstringset(x, remove_gaps = FALSE)
}
\arguments{
\item{x}{a DNABin object}

\item{remove_gaps}{Whether gaps should be removed}
}
\value{

}
\description{
Convert DNABin to DNAStringSet
}
