% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{.findExecutable}
\alias{.findExecutable}
\title{Find executable}
\usage{
.findExecutable(exe, quiet = FALSE)
}
\arguments{
\item{exe}{the name of the executable to find}

\item{quiet}{Whether errors should be printed to console}
}
\value{

}
\description{
Find executable
}
