% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.hex}
\alias{.hex}
\title{convert alphanumerics to hexadecimal
internal taxreturn function}
\usage{
.hex(y)
}
\arguments{
\item{y}{a character string}
}
\value{

}
\description{
convert alphanumerics to hexadecimal
internal taxreturn function
}
