% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.ishex}
\alias{.ishex}
\title{Check whether a string is hexadecimal
internal taxreturn function}
\usage{
.ishex(y)
}
\arguments{
\item{y}{a character string}
}
\value{

}
\description{
Check whether a string is hexadecimal
internal taxreturn function
}
