% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.unhex}
\alias{.unhex}
\title{convert hexadecimal strings to alphanumeric
internal taxreturn function}
\usage{
.unhex(y)
}
\arguments{
\item{y}{a character string}
}
\description{
convert hexadecimal strings to alphanumeric
internal taxreturn function
}
