% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_utilities.R
\name{fastqc}
\alias{fastqc}
\title{Run FastQC Tool}
\usage{
fastqc(fq.dir, qc.dir = NULL, threads = 1, fastqc.path = "bin/FastQC/fastqc")
}
\arguments{
\item{fq.dir}{(Required) Default is the current working directory.
path to the directory containing fastq files.}

\item{qc.dir}{(Optional) path to the FastQC result directory. If NULL, a directory
named FASTQC is created in the current working directory.}

\item{threads}{(Optional) the number of threads to be used. Default is 1.}

\item{fastqc.path}{(Required) path to fastqc program}
}
\value{
A new directory containing the fastqc reports for each sample
}
\description{
Run FastQC Tool on windows or linux
}
\examples{
\dontrun{
# Run FastQC: generates a QC directory
fastqc(fq.dir)
}
}
