% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_utilities.R
\name{fastqc_install}
\alias{fastqc_install}
\title{fastqc_install}
\usage{
fastqc_install(url, dest.dir = "bin", force = FALSE)
}
\arguments{
\item{url}{(Optional) Default will search for the latest version
URL to retrieve fastQC from.}

\item{dest.dir}{(Optional)  Default "bin"
Directory to install fastQC within.}
}
\value{

}
\description{
fastqc_install
}
