% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{fetch_bold}
\alias{fetch_bold}
\title{Fetch sequences from BOLD}
\usage{
fetch_bold(
  x,
  marker = "COI-5P",
  output = "gb-binom",
  quiet = FALSE,
  db = NULL,
  retry_attempt = 3,
  retry_wait = 5
)
}
\arguments{
\item{x}{A taxon name or vector of taxa to download sequences for}

\item{marker}{the barcode marker used as a search term for the database}

\item{output}{the output format for the taxonomy in fasta headers.
Options include "h" for full heirarchial taxonomy (Accession;Domain;Phylum;Class;Order;Family;Genus;Species),
"binom" for just genus species binomials (Accession;Genus_species),
"bold" for BOLD taxonomic ID only (Accession;BoldTaxID),
"gb" for genbank taxonomic ID (Accession|GBTaxID),
"gb-binom" which outputs genbank taxonomic ID's and Genus species binomials, translating BOLD taxonomic ID's to genbank in the process (Accession|GBTaxID;Genus_species)
or "standard" which outputs the default format for each database. For genbank this is the description field, and for bold this is \verb{sampleid|species name|markercode|genbankid}}

\item{quiet}{Whether progress should be printed to the console.}

\item{db}{(Optional) a database file generated using \code{taxreturn::get_ncbi_taxonomy()} or \code{taxreturn::get_ott_taxonomy()}}

\item{retry_attempt}{The number of query attempts in case of query failure due to poor internet connection.}

\item{retry_wait}{How long to wait between query attempts}
}
\value{

}
\description{
Fetch sequences from BOLD
}
