% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{fetch_seqs2}
\alias{fetch_seqs2}
\title{Title}
\usage{
fetch_seqs2(
  x,
  database,
  marker = NULL,
  downstream = FALSE,
  output = "gb-binom",
  min_length = 1,
  max_length = 2000,
  subsample = FALSE,
  chunk_size = NULL,
  db = NULL,
  multithread = FALSE,
  quiet = FALSE,
  progress = FALSE
)
}
\arguments{
\item{progress}{}
}
\value{

}
\description{
Title
}
