% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_taxa (Alexander’s MacBook Pro's
%   conflicted copy 2019-06-14).R, R/fetch_taxa.R
\name{fetchSeqs}
\alias{fetchSeqs}
\title{Fetchseqs wrapper function}
\usage{
fetchSeqs(x, database, marker = "COI", downstream = FALSE,
  downto = "family", quiet = TRUE, output = "h", minlength = 1,
  maxlength = 2000, dir = NULL, compress = FALSE, cores = 1)

fetchSeqs(x, database, marker = "COI", downstream = FALSE,
  downto = "family", quiet = TRUE, output = "h", minlength = 1,
  maxlength = 2000, dir = NULL, compress = FALSE, cores = 1)
}
\arguments{
\item{cores}{}

\item{cores}{}
}
\description{
Fetchseqs wrapper function

Fetchseqs wrapper function
}
