% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{filter_by_tax}
\alias{filter_by_tax}
\title{Filter sequences by taxonomy}
\usage{
filter_by_tax(
  x,
  filt_rank,
  filt_value,
  ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{x}{A DNAbin or DNAString with heirarchial taxonomy}

\item{filt_rank}{The taxonomic rank to subset at i.e. 'Class'}

\item{filt_value}{A taxonomic name or vector of taxonomic names to subset filt_rank at at i.e. c('Insecta', 'Arachnida')}

\item{ranks}{The taxonomic ranks currently assigned to the sequences}
}
\description{
Filter sequences by taxonomy
}
