% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{filter_infraspecifc}
\alias{filter_infraspecifc}
\title{Filter infraspecific taxonomic labels}
\usage{
filter_infraspecifc(x, db, quiet = FALSE)
}
\arguments{
\item{x}{A DNAbin, DNAStringSet, taxonomy headers formnatted Acc|taxid;taxonomy, or vector of taxids}

\item{db}{an OTT taxonomic database}

\item{quiet}{Whether progress should be printed to the console.}
}
\value{

}
\description{
Filter infraspecific taxonomic labels
}
