% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{filter_unplaced}
\alias{filter_unplaced}
\title{Filter unplaced taxonomic labels}
\usage{
filter_unplaced(x, db, quiet = FALSE)
}
\arguments{
\item{x}{A DNAbin, DNAStringSet, taxonomy headers formnatted Acc|taxid;taxonomy, or vector of taxids}

\item{db}{an OTT taxonomic database}

\item{quiet}{Whether progress should be printed to the console.}
}
\description{
Filter flags indicating unplaced taxa in the taxonomic tree. These include:
incertae_sedis
major_rank_conflict
unplaced
environmental
inconsistent
extinct
hidden
hybrid
not_otu
viral
barren
See: https://github.com/OpenTreeOfLife/reference-taxonomy/blob/master/doc/taxon-flags.md for more info
}
