% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{find_mixed_clusters}
\alias{find_mixed_clusters}
\title{Find mixed clusters}
\usage{
find_mixed_clusters(
  x,
  db,
  level = "order",
  confidence = 0.8,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Find mixed clusters
}
