% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_taxa (Alexander’s MacBook Pro's
%   conflicted copy 2019-06-14).R, R/fetch_taxa.R
\name{gbSearch}
\alias{gbSearch}
\title{Genbank search function}
\usage{
gbSearch(x, marker = "COI", quiet = FALSE, output = "h",
  minlength = 1, maxlength = 2000, file = NULL, compress = FALSE,
  dir = NULL)

gbSearch(x, marker = "COI", quiet = FALSE, output = "h",
  minlength = 1, maxlength = 2000, file = NULL, compress = FALSE,
  dir = NULL)
}
\arguments{
\item{dir}{}

\item{dir}{}
}
\description{
Genbank search function

Genbank search function
}
