% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{gbUpdate}
\alias{gbUpdate}
\title{GbUpdate}
\usage{
gbUpdate(
  x,
  fasta,
  database = "nuccore",
  marker = c("COI[GENE]", "CO1[GENE]", "COX1[GENE]"),
  quiet = FALSE,
  output = "h",
  suffix = "updates",
  min_length = 1,
  max_length = 2000,
  chunk_size = 300,
  out_dir = NULL,
  compress = FALSE,
  force = FALSE,
  multithread = FALSE,
  progress = FALSE
)
}
\arguments{
\item{x}{A taxon name or vector of taxa to download sequences for}

\item{fasta}{A fasta file or list of fasta files to check for existing sequence accessions}

\item{marker}{The barcode marker used as a search term for the database. If this is set to "mitochondria" it will download full mitochondrial genomes.}

\item{quiet}{(Optional) Print text output}

\item{output}{The output format for the taxonomy in fasta headers.
Options include "h" for full heirarchial taxonomy (SeqID;Domain;Phylum;Class;Order;Family;Genus;Species),
"binom" for just genus species binomials (SeqID;Genus Species),
"bold" for BOLD taxonomic ID only (SeqID;BoldTaxID),
"gb" for genbank taxonomic ID (SeqID;GBTaxID),
or "gb-binom" which outputs Genus species binomials, as well as genbank taxonomic ID's, and translates all BOLD taxonomic ID's to genbank taxonomic ID's in the process}

\item{suffix}{The suffix to add to newly downloaded files. Defaults to 'updates'}

\item{min_length}{The minimum length of sequences to download}

\item{max_length}{The maximum length of sequences to download}

\item{chunk_size}{The size of the chunked searches to conduct.
Warning, chunk sizes over 300 can be too big for the NCBI servers.}

\item{out_dir}{Output directory to write fasta files to}

\item{compress}{Option to compress output fasta files using gzip}

\item{force}{Option ot overwright files if they already exist}

\item{multithread}{Whether multithreading should be used, if TRUE the number of cores will be automatically detected, or provided a numeric vector to manually set the number of cores to use}

\item{progress}{Whether a progress bar should be shown. This reduces speed and is false by default.}
}
\value{

}
\description{
GbUpdate
}
