% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_binding_position}
\alias{get_binding_position}
\title{Get primer binding position}
\usage{
get_binding_position(primer, model, tryrc = TRUE, quiet = FALSE,
  minscore = 10, ...)
}
\arguments{
\item{primer}{A character string, DNAbin object or an object coercible to DNAbin}

\item{model}{A profile hidden Markov model (a "PHMM" object) generated by the aphid R package to align the sequences to.}

\item{tryrc}{Whether the reverse complement should also be aligned. The highest scoring complement is chosen.}

\item{quiet}{Whether progress should be printed to the console.}

\item{minscore}{Minimum score for the viterbi alignment.}

\item{...}{aditional arguments to be passed to "Viterbi"}
}
\description{
Get primer binding position
}
