% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{get_mixed_clusters}
\alias{get_mixed_clusters}
\title{Get mixed clusters}
\usage{
get_mixed_clusters(
  x,
  db,
  rank = "order",
  threshold = 0.97,
  confidence = 0.8,
  return_all_taxa = FALSE,
  nstart = 20,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{A DNAbin list object whose names include NCBItaxonomic identification numbers.}

\item{db}{A taxonomic database from \code{get_ncbi_lineage} or \code{get_ott_lineage}}

\item{rank}{The taxonomic rank to check clusters at, accepts a character such as "order", or vector of characters such as c("species", "genus").
If "all", the clusters will be checked at all taxonomic ranks available.}

\item{threshold}{numeric between 0 and 1 giving the OTU identity cutoff for clustering. Defaults to 0.97.}

\item{confidence}{The minimum confidence value for a mixed cluster to be flagged. For example, if confidence = 0.8 (the default value)
a cluster will only be flagged if the taxonomy of a sequence within the cluster differs from at least four other independent sequences in its cluster.
@param nstart how many random sets should be chosen for \code{kmeans}, It is recommended to set the value of nstart to at least 20.
While this can increase computation time, it can improve clustering accuracy considerably.}

\item{return_all_taxa}{Option to return all unique taxa within mixed clusters as a list column.}

\item{quiet}{logical indicating whether progress should be printed to the console.}

\item{...}{further arguments to pass to kmer::otu.}
}
\value{

}
\description{
Get mixed clusters
}
\examples{
\dontrun{
seqs <- insect::readFASTA("test.fa.gz")

# NCBI taxonomy
mixed <- get_mixed_clusters(seqs, db, rank="species", threshold=0.99, confidence=0.8, quiet=FALSE)

# OTT taxonomy
seqs <- map_to_ott(seqs, dir="ott3.2", from="ncbi", resolve_synonyms=TRUE, filter_bads=TRUE, remove_na = TRUE, quiet=FALSE)
mixed <- get_mixed_clusters(seqs, db, rank="species", threshold=0.99, confidence=0.6, quiet=FALSE)
}
}
