% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_ncbi_lineage}
\alias{get_ncbi_lineage}
\title{Get lineage}
\usage{
get_ncbi_lineage(x, db)
}
\arguments{
\item{x}{A DNAbin or DNAStringSet object with names in format \verb{accession|tax_id;Genus species}}

\item{db}{a database file generated using \code{taxreturn::get_ncbi_taxonomy()}}
}
\value{

}
\description{
Get lineage
}
