% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_ncbi_synonyms}
\alias{get_ncbi_synonyms}
\title{Get NCBI synonyms}
\usage{
get_ncbi_synonyms(dir = NULL, quiet = FALSE)
}
\arguments{
\item{dir}{A directory containing the NCBI taxonomy that was downloaded using get_ncbi_lineage}

\item{quiet}{}
}
\value{

}
\description{
Get NCBI synonyms
}
