% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_ncbi_taxonomy}
\alias{get_ncbi_taxonomy}
\title{Download NCBI taxdump}
\usage{
get_ncbi_taxonomy(dest_dir, include_synonyms = TRUE, force = FALSE)
}
\arguments{
\item{dest_dir}{A directory to save the downloaded ncbi taxdump files. If empty a new folder in the working directory will be created}

\item{include_synonyms}{Whether synonyms should be included in the database}

\item{force}{Whether already downloaded files should be overwrittenget}
}
\value{

}
\description{
Download NCBI taxdump
}
