% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{get_ott_taxonomy}
\alias{get_ott_taxonomy}
\title{get_ott_taxonomy}
\usage{
get_ott_taxonomy(dir = NULL, quiet = FALSE, filter_bads = TRUE)
}
\arguments{
\item{dir}{a directory containing the open tree of life taxonomy files obtained from the  \code{download_ott_taxonomy} function}

\item{quiet}{Whether progress should be printed to console}

\item{filter_bads}{Whether to filter 'bad' entries. These include
incertae_sedis
major_rank_conflict
unplaced
environmental
inconsistent
extinct
hidden
hybrid
not_otu
viral
barren
See: https://github.com/OpenTreeOfLife/reference-taxonomy/blob/master/doc/taxon-flags.md for more info}
}
\value{

}
\description{
get_ott_taxonomy
}
