% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{get_ranked_lineage}
\alias{get_ranked_lineage}
\title{Download NCBI taxdump}
\usage{
get_ranked_lineage(db = "NCBI", synonyms = TRUE, force = FALSE)
}
\arguments{
\item{synonyms}{}
}
\value{

}
\description{
Download NCBI taxdump
}
