% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{get_reading_frame}
\alias{get_reading_frame}
\title{Get Reading frame of sequences}
\usage{
get_reading_frame(
  x,
  genetic_code = NULL,
  forward = TRUE,
  reverse = FALSE,
  resolve_draws = "majority"
)
}
\arguments{
\item{x}{Sequences in DNAStringset or DNAbin format}

\item{genetic_code}{A genetic code for the Amino acid translation. set to 'SGC4' for Invertebrate mitochondrial or see all known codes at Biostrings::GENETIC_CODE_TABLE}

\item{forward}{whether the forward complement should be returned}

\item{reverse}{Whether the reverse complemement should be returned}

\item{resolve_draws}{How draws should be resolved when multiple possible frames produce sequences with no stop codons.}
}
\value{

}
\description{
Get Reading frame of sequences
}
