% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_subalignment}
\alias{get_subalignment}
\title{Get subalignment}
\usage{
get_subalignment(
  x,
  model,
  tryrc = FALSE,
  quiet = FALSE,
  check_indels = TRUE,
  minscore = 10,
  omit.endgaps = FALSE,
  cores = 1,
  ...
)
}
\arguments{
\item{x}{A DNAbin object or an object coercible to DNAbin}

\item{model}{A profile hidden Markov model (a "PHMM" object) generated by the aphid R package to align the sequences to.}

\item{tryrc}{Whether the reverse complement should also be aligned}

\item{quiet}{Whether progress should be printed to the console.}

\item{check_indels}{Check that indels are multiples of 3, recommended for coding sequences such as COI}

\item{minscore}{Minimum score for the viterbi alignment}

\item{omit.endgaps}{Should gap characters at each end of the sequences be ignored when deriving the transition probabilities of the model? Defaults to FALSE. Set to TRUE if x is not a strict global alignment (i.e. if the alignment contains partial sequences with missing sections represented with gap characters).}

\item{cores}{Integer giving the number of CPUs to parallelize the operation over. Defaults to 1, and reverts to 1 if x is not a list. This argument may alternatively be a 'cluster' object, in which case it is the user's responsibility to close the socket connection at the conclusion of the operation, for example by running parallel::stopCluster(cores). The string 'autodetect' is also accepted, in which case the maximum number of cores to use is one less than the total number of cores available. Note that in this case there may be a tradeoff in terms of speed depending on the number and size of sequences to be aligned, due to the extra time required to initialize the cluster.}

\item{...}{aditional arguments to be passed to "Viterbi"}
}
\value{

}
\description{
Aligns a DNABin to a reference PHMM model, and returns the optimal path
}
