% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{graft_tips}
\alias{graft_tips}
\title{Graft new tips onto an existing tree}
\usage{
graft_tips(tree, mapping, branch_length = 0)
}
\arguments{
\item{tree}{A phylo object}

\item{mapping}{A mapping data frame containing two columns.
The left column must be identical to the tip labels of the phylogeny, the right column contains new tips to be grafted.}

\item{branch_length}{A numeric giving the branch lengths for the new polytomic tips that are added.}
}
\description{
Graft new tips onto an existing tree
}
