% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hex2acc}
\alias{hex2acc}
\title{Hexadecimal coding to accession numberD}
\usage{
hex2acc(x, force = FALSE)
}
\arguments{
\item{x}{A DNAbin or DNAStringSet with names formatted Accession|taxid;taxonomy.
Or a character vector with names formatted Accession|taxid;taxonomy
Or a character vector of accessions}

\item{force}{override checks if string is already hexadecimal}
}
\value{

}
\description{
Hexadecimal coding is useful for ensuring that accession numbers dont contain characters that are considered illegal for newick strings
}
