% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{kmer_screen}
\alias{kmer_screen}
\title{Kmer screening function}
\usage{
kmer_screen(seqs, model, threshold = 0.3, k = 5, quiet = FALSE)
}
\arguments{
\item{seqs}{A DNAbin object}

\item{model}{A Profile Hidden Markov model ("PHMM" object) generated by \code{aphid::derivePHMM} to screen the sequences against.
An already derived model of COI can be loaded using \code{data("model", package="taxreturn")}}

\item{threshold}{The maximum kmer distance allowed from the reference model before a sequence is discarded.}

\item{k}{The k-mer size to be used for calculating the distance matrix, parsed to kmer::mbed. Note that high values of k may be slow to compute and use a lot of memory due to the large numbers of calculations required, particularly when the residue alphabet is also large.}

\item{quiet}{Whether progress should be printed to the console.}
}
\description{
Kmer screening function
}
