% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{lca_probs}
\alias{lca_probs}
\title{Probabilitys of sharing a rank as a function of sequence identity}
\usage{
lca_probs(
  x,
  method = "mbed",
  k = 5,
  nstart = 20,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  delim = ";"
)
}
\arguments{
\item{x}{a DNAbin object or an object coercible to DNAbin}

\item{method}{The distance matrix computation method to use,
accepts "mbed" which computes a distance matrix from each sequence to a subset of 'seed' sequences using the method outlined in Blacksheilds et al (2010).
This scales well to big datasets, alternatively "kdist" computes the full n * n distance matrix.}

\item{k}{integer giving the k-mer size used to generate the input matrix for k-means clustering.}

\item{nstart}{value passed to  nstart passed to kmeans. Higher increases computation time but can improve clustering accuracy considerably.}

\item{ranks}{The taxonomic ranks currently assigned to the names}

\item{delim}{The delimiter used between ranks}
}
\value{

}
\description{
Probabilitys of sharing a rank as a function of sequence identity
}
