% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Make blast Database (DEPRECATED)}
\usage{
makeblastdb(file, dbtype = "nucl", args = NULL, quiet = FALSE)
}
\arguments{
\item{file}{(Required) A fasta file to create a database from.}

\item{dbtype}{(Optional) Molecule type of database, accepts "nucl" for nucleotide or "prot" for protein.}

\item{args}{(Optional) Extra arguments passed to BLAST}

\item{quiet}{(Optional) Whether progress should be printed to console, default is FALSE}
}
\description{
Make blast Database (DEPRECATED)
}
