% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{map_to_model}
\alias{map_to_model}
\title{This function maps sequences to a Profile Hidden Markov Model (PHMM) in order to retain only the target loci}
\usage{
map_to_model(
  x,
  model,
  minscore = 100,
  shave = TRUE,
  check_indels = FALSE,
  extra = NA,
  maxNs = Inf,
  multithread = FALSE,
  quiet = FALSE,
  progress = FALSE,
  ...
)
}
\arguments{
\item{x}{A DNAbin or DNAStringset object}

\item{model}{A Profile Hidden Markov model ("PHMM" object) generated by \code{aphid::derivePHMM} to align the sequences to.
An already derived model of COI can be loaded using \code{data("model", package="taxreturn")}}

\item{minscore}{The minimum specificity (log-odds score for the optimal alignment) between the query sequence and the PHMM model for the sequence to be retained in the output object.
see \code{?aphid::Viterbi} for more information about the alignment process.}

\item{shave}{Whether bases that are outside (to the left or right) of the PHMM object should be removed from sequences in the output. Default is TRUE.}

\item{check_indels}{Whether sequences with insertions or deletions which arent in multiples of 3 should be removed from output. Useful for coding loci such as COI but should not be used for non-coding loci. Default is FALSE.}

\item{extra}{How to handle insertions which were not part of the PHMM model. 'drop' will truncate all sequences to the shortest alignment length, while 'fill' will use gaps to pad all sequences out to the longest alignment length.}

\item{maxNs}{The max number of ambiguous N bases allowed before a sequence is removed.}

\item{multithread}{Whether multithreading should be used, if TRUE the number of cores will be automatically detected (Maximum available cores - 1), or provide a numeric vector to manually set the number of cores to use. Default is FALSE (single thread)}

\item{quiet}{Whether progress should be printed to the console. Note that this will add additional runtime.}

\item{...}{Any further arguments to be passed to \code{aphid::Viterbi}}
}
\value{

}
\description{
This function maps sequences to a Profile Hidden Markov Model (PHMM) in order to retain only the target loci
}
