% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model}
\alias{model}
\title{COI phmm model}
\format{
A Profile hidden Markov model (object class: 'PHMM') with 658 internal modules emitting 4 unique residues
}
\usage{
model
}
\description{
A Profile Hidden Markov Model of the COI barcode region
}
\keyword{datasets}
