% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{ncbi_taxid}
\alias{ncbi_taxid}
\title{Get ncbi taxid's for a taxon name}
\usage{
ncbi_taxid(x, db = NULL)
}
\arguments{
\item{x}{A DNAbin or DNAStringSet object with names in format \verb{accession|tax_id;Genus species}}

\item{db}{a database file generated using \code{taxreturn::get_ncbi_taxonomy()}}
}
\description{
Get ncbi taxid's for a taxon name
}
