% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pad_alignment}
\alias{pad_alignment}
\title{Pad alignment}
\usage{
pad_alignment(
  x,
  model,
  pad = "-",
  tryrc = FALSE,
  check_indels = TRUE,
  min_score = 10,
  omit_endgaps = FALSE,
  multithread = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{A DNAbin object or an object coercible to DNAbin}

\item{model}{A profile hidden Markov model (a "PHMM" object) generated by the aphid R package to align the sequences to.}

\item{pad}{The character used to pad the gaps}

\item{tryrc}{Whether the reverse complement should also be aligned}

\item{check_indels}{Check that indels are multiples of 3, recommended for coding sequences such as COI}

\item{min_score}{Minimum score for the viterbi alignment}

\item{omit_endgaps}{Should gap characters at each end of the sequences be ignored when deriving the transition probabilities of the model? Defaults to FALSE. Set to TRUE if x is not a strict global alignment (i.e. if the alignment contains partial sequences with missing sections represented with gap characters).}

\item{multithread}{Whether multithreading should be used, if TRUE the number of cores will be automatically detected, or provided a numeric vector to manually set the number of cores to use}

\item{quiet}{Whether progress should be printed to the console.}

\item{...}{aditional arguments to be passed to "Viterbi"}
}
\description{
Aligns a DNABin to a reference PHMM model, and pads any gaps between the query and reference
}
