% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_aqhist}
\alias{parse_aqhist}
\title{Parse average quality histograms from bbtrim}
\usage{
parse_aqhist(dir)
}
\arguments{
\item{dir}{(Required) a directory containing average quality (aqhist) files from bbtrim}
}
\value{
a tidy data frame
}
\description{
Parse average quality histograms from bbtrim
}
