% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_bbdemux}
\alias{parse_bbdemux}
\title{Parse appended bbdemux logs into tidy format}
\usage{
parse_bbdemux(x)
}
\arguments{
\item{x}{(Required) an appended bbdemux log file generated by bbdemux}
}
\value{
Returns a tidy data frame
}
\description{
Parse appended bbdemux logs into tidy format
}
