% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_bbtrim}
\alias{parse_bbtrim}
\title{Parse appended bbtrim logs into tidy format}
\usage{
parse_bbtrim(x)
}
\arguments{
\item{x}{(Required) an appended bbtrim log file generated by bbtrim}
}
\value{
Returns a tidy data frame
}
\description{
Parse appended bbtrim logs into tidy format
}
