% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_bhist}
\alias{parse_bhist}
\title{parse base frequency histograms from bbtrim}
\usage{
parse_bhist(dir)
}
\arguments{
\item{dir}{(Required) a directory containing base frequency (bhist) files from bbtrim}
}
\value{
a tidy data frame
}
\description{
parse base frequency histograms from bbtrim
}
