% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_gchist}
\alias{parse_gchist}
\title{Parse GC content histograms from bbtrim}
\usage{
parse_gchist(dir)
}
\arguments{
\item{dir}{(Required) a directory containing average quality (aqhist) files from bbtrim}
}
\value{
a tidy data frame
}
\description{
Parse GC content histograms from bbtrim
}
