% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_lhist}
\alias{parse_lhist}
\title{Parse sequence length histograms from bbtrim}
\usage{
parse_lhist(dir)
}
\arguments{
\item{dir}{(Required) a directory containing length (lhist) files from bbtrim}
}
\value{
a tidy data frame
}
\description{
Parse sequence length histograms from bbtrim
}
