% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{parse_ott_lineage}
\alias{parse_ott_lineage}
\title{Parse OTT taxonomy to ranked_lineage format}
\usage{
parse_ott_lineage(dir = NULL)
}
\arguments{
\item{dir}{a directory containing the open tree of life taxonomy files obtained from the  \code{download_ott_taxonomy} function}
}
\value{

}
\description{
Parse OTT taxonomy to ranked_lineage format
}
