% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{parse_ott_synonyms}
\alias{parse_ott_synonyms}
\title{parse the open tree of life synonyms file}
\usage{
parse_ott_synonyms(dir = NULL, quiet = FALSE)
}
\arguments{
\item{dir}{a directory containing the open tree of life taxonomy files obtained from the  \code{download_ott_taxonomy} function}

\item{quiet}{Whether progress should be printed to console}
}
\description{
parse the open tree of life synonyms file
}
