% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{parse_qhist}
\alias{parse_qhist}
\title{Parse quality score histograms from bbtrim}
\usage{
parse_qhist(dir)
}
\arguments{
\item{dir}{(Required) a directory containing quality score (qhist) files from bbtrim}
}
\value{
a tidy data frame
}
\description{
Parse quality score histograms from bbtrim
}
