% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{propagate_tax}
\alias{propagate_tax}
\title{Propagate taxonomy (DEPRECATED)}
\usage{
propagate_tax(tax, from = "Family")
}
\arguments{
\item{tax}{A taxonomy table}

\item{from}{The taxonomic rank to propagate below}
}
\description{
Propagate taxonomy (DEPRECATED)
}
