% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{proportions}
\alias{proportions}
\title{Convert phyloseq table to proportions}
\usage{
proportions(x, thresh = NA, na_rm = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Convert phyloseq table to proportions
}
