% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{qualMaxEEplot}
\alias{qualMaxEEplot}
\title{QC Max EE plot}
\usage{
qualMaxEEplot(fastq.r1, fastq.r2, name1 = "Fastq_R1",
  name2 = "Fastq_R2")
}
\arguments{
\item{name2}{}
}
\description{
QC Max EE plot
}
