% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{random_seq}
\alias{random_seq}
\title{Generate random sequences}
\usage{
random_seq(n, length, alphabet = c("A", "G", "T", "C"))
}
\arguments{
\item{n}{The number of sequences to generate}

\item{length}{The length of the generated sequences}

\item{alphabet}{The DNA alphabet to draw from, default is A,G,C,T}
}
\description{
Generate random sequences
}
