% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{read_bold_chunk}
\alias{read_bold_chunk}
\title{read bold chunk}
\usage{
read_bold_chunk(taxon, quiet = FALSE, retry_attempt = 3, retry_wait = 5)
}
\arguments{
\item{taxon}{A taxon name to download sequences for}

\item{quiet}{Whether progress should be printed to console.}

\item{retry_attempt}{The number of query attempts in case of query failure due to poor internet connection.}

\item{retry_wait}{How long to wait between query attempts}

\item{gid}{a vector of GenBank ID's}
}
\value{

}
\description{
read bold chunk
}
