% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{read_genbank_chunk}
\alias{read_genbank_chunk}
\title{Read genbank chunk}
\usage{
read_genbank_chunk(gid, quiet = FALSE, retry_attempt = 3, retry_wait = 5)
}
\arguments{
\item{gid}{a vector of GenBank ID's}

\item{quiet}{Whether progress should be printed to console.}

\item{retry_attempt}{The number of query attempts in case of query failure due to poor internet connection.}

\item{retry_wait}{How long to wait between query attempts}
}
\value{

}
\description{
Read genbank chunk
}
