% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_dada2_gen}
\alias{reformat_dada2_gen}
\title{Reformat DADA2 Genus}
\usage{
reformat_dada2_gen(
  x,
  db = NULL,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus"),
  quiet = FALSE
)
}
\arguments{
\item{x}{A DNAbin or DNAStringSet object with names in format \verb{accession|tax_id;Genus species}}

\item{db}{A database file generated by \code{get_ncbi_taxonomy} or \code{get_ott_lineage}. Required if get_lineage is TRUE}

\item{ranks}{The taxonomic ranks to include in output. default is kingdom;phylum;class;order;family;genus}

\item{quiet}{Whether progress should be printed to the console.}
}
\description{
Reformat DADA2 Genus
}
