% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_heirarchy}
\alias{reformat_heirarchy}
\title{Reformat heirarchy}
\usage{
reformat_heirarchy(x, db = NULL, quiet = FALSE, ranks = NULL,
  sppsep = "_", force = FALSE)
}
\arguments{
\item{ranks}{}
}
\description{
Reformat heirarchy
}
