% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_hierarchy}
\alias{reformat_hierarchy}
\title{Reformat hierarchy}
\usage{
reformat_hierarchy(
  x,
  db = NULL,
  quiet = FALSE,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{ranks}{}
}
\value{

}
\description{
Reformat hierarchy
}
