% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{resolve_gbif}
\alias{resolve_gbif}
\title{Resolve GBIF}
\usage{
resolve_gbif(
  x,
  subspecies = TRUE,
  higherrank = TRUE,
  verbose = FALSE,
  fuzzy = TRUE,
  conf_threshold = 90,
  resolve_taxonomy = TRUE
)
}
\arguments{
\item{resolve_taxonomy}{}
}
\value{

}
\description{
Resolve GBIF
}
