% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{resolve_synonyms}
\alias{resolve_synonyms}
\title{Resolve taxonomic synonyms}
\usage{
resolve_synonyms(x, subspecies = FALSE, quiet = TRUE,
  missing = "ignore", higherrank = FALSE, fuzzy = TRUE)
}
\arguments{
\item{fuzzy}{}
}
\description{
Resolve taxonomic synonyms
}
