% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{resolve_synonyms_ncbi}
\alias{resolve_synonyms_ncbi}
\title{resolve_synonyms_ncbi}
\usage{
resolve_synonyms_ncbi(x, dir = NULL, quiet = FALSE, ...)
}
\arguments{
\item{x}{A DNAbin or DNAStringSet Object}

\item{quiet}{Whether progress should be printed to console}
}
\value{

}
\description{
resolve_synonyms_ncbi
}
