% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{resolve_taxonomy}
\alias{resolve_taxonomy}
\title{Resolve taxonomic synonyms}
\usage{
resolve_taxonomy(x, subspecies = FALSE, quiet = TRUE,
  missing = "ignore", higherrank = FALSE, fuzzy = TRUE)
}
\arguments{
\item{fuzzy}{}
}
\description{
This function takes a DNAbin object, or a list of species and uses the Global Biodiversity Information Facility (GBIF) to resolve taxonomic synonyms
}
