% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{search_genbank}
\alias{search_genbank}
\title{all \verb{[filter]} 	Retrieves everthing
Specified \verb{[property]} 	Formal binomial and trinomial
at or below species level \verb{[property]}
family \verb{[rank]} 	Rank-based query
taxonomy genome \verb{[filter]} 	Taxa with a direct link to a genome sequence
2009/10/21:2020 \verb{[date]} 	Date-bounded query
mammalia \verb{[subtree]} 	All taxa within the Mammalia
extinct \verb{[property]} 	Extinct organisms
Terminal \verb{[property]} 	Terminal nodes in the tree
loprovencyclife \verb{[filter]} 	Entries with LinkOut links to the Encyclopedia of Life}
\usage{
search_genbank(
  x,
  database = "nuccore",
  marker = c("COI[GENE]", "CO1[GENE]", "COX1[GENE]"),
  quiet = FALSE,
  output = "h",
  min_length = 1,
  max_length = 2000,
  chunk_size = NULL,
  out_dir = NULL,
  compress = FALSE,
  force = FALSE
)
}
\arguments{
\item{x}{A taxon name or vector of taxa to download sequences for}

\item{database}{The database to download from. For NCBI GenBank this currently onlt accepts the arguments 'nuccore' or 'genbank' which is an alias for nuccore.}

\item{marker}{The barcode marker used as a search term for the database.
If this is set to "mitochondria" or "mitochondrion" it will download full mitochondrial genomes. If set to "genome" it will download entire genomes only.}

\item{quiet}{Whether progress should be printed to the console.}

\item{output}{The output format for the taxonomy in fasta headers.
Options include "h" for full heirarchial taxonomy (SeqID;Domain;Phylum;Class;Order;Family;Genus;Species),
"binom" for just genus species binomials (SeqID;Genus Species),
"bold" for BOLD taxonomic ID only (SeqID;BoldTaxID),
"gb" for genbank taxonomic ID (SeqID;GBTaxID),
"gb-binom" which outputs Genus species binomials, as well as genbank taxonomic ID's, and translates all BOLD taxonomic ID's to genbank taxonomic ID's in the process
or "standard" which outputs the default format for each database. For genbank this is \verb{Accession Sequence definition}}

\item{min_length}{The minimum length of sequences to download}

\item{max_length}{The maximum length of sequences to download}

\item{chunk_size}{Split up the query into chunks of this size to avoid overloading API servers.
if left NULL, the default will be 10,000 for regular queries, 1,000 if marker is "mitochondria", and 1 if marker is "genome"}

\item{out_dir}{Output directory to write fasta files to}

\item{compress}{Option to compress output fasta files using gzip}

\item{force}{Option ot overwright files if they already exist}
}
\value{

}
\description{
Genbank search function
}
\details{

}
