% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_multithread}
\alias{setup_multithread}
\title{Setup multithreading}
\usage{
setup_multithread(multithread, quiet = FALSE)
}
\arguments{
\item{multithread}{Number of cores}

\item{quiet}{Whether progress should be printed to console}
}
\description{
Setup multithreading
}
