% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{setup_para}
\alias{setup_para}
\title{Setup parallel}
\usage{
setup_para(multithread, quiet = FALSE)
}
\arguments{
\item{multithread}{Number of cores}

\item{quiet}{Whether progress should be printed to console}
}
\value{

}
\description{
Setup parallel
}
