% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{split_bold_query}
\alias{split_bold_query}
\title{Split bold query
This function recursively splits a bold taxonomic query until the amount of records returned is under chunksize}
\usage{
split_bold_query(x, chunksize = 1e+05, quiet = FALSE)
}
\arguments{
\item{x}{The input taxonomic query}

\item{chunksize}{The maximum amount of records to return per query}

\item{quiet}{(Optional) Print text output}
}
\value{

}
\description{
Split bold query
This function recursively splits a bold taxonomic query until the amount of records returned is under chunksize
}
