% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{subset_model}
\alias{subset_model}
\title{Subset PHMM
Modified from Cameron Nugent's COIL package https://github.com/CNuge/coil/blob/master/R/subset_PHMM.r}
\usage{
subset_model(
  x,
  primers = NULL,
  positions = NULL,
  trimprimers = FALSE,
  min_score = 10,
  quiet = FALSE
)
}
\arguments{
\item{x}{A profile hidden Markov model (a "PHMM" object) generated by the aphid R package.}

\item{primers}{A character vector of length 2 contaning the forward and reverse primer sequences.}

\item{positions}{A numeric vector of length 2 containing the start and end position for model trimming. Used only if primers = NULL.}

\item{trimprimers}{logical indicating whether the primer-binding sites should be removed from the sequences in the returned model. Used only if positions = NULL.}

\item{min_score}{Minimum score for the viterbi alignment.}

\item{quiet}{Whether to print progress to console.}
}
\value{

}
\description{
Subset PHMM
Modified from Cameron Nugent's COIL package https://github.com/CNuge/coil/blob/master/R/subset_PHMM.r
}
