% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{subset_tax}
\alias{subset_tax}
\title{Subset sequences by taxonomy}
\usage{
subset_tax(
  x,
  filtrank,
  filtvalue,
  ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{x}{A DNAbin or DNAString with heirarchial taxonomy}

\item{filtrank}{The taxonomic rank to subset at i.e. 'Class'}

\item{filtvalue}{The taxonomic name to subset at i.e. 'Insecta'}

\item{ranks}{The taxonomic ranks currently assigned to the sequences}
}
\value{

}
\description{
Subset sequences by taxonomy
}
