% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{summarise_fasta}
\alias{summarise_fasta}
\title{summarise_fasta}
\usage{
summarise_fasta(x, label = NULL, origin = NULL)
}
\arguments{
\item{x}{The location of a fasta file or gzipped fasta file.}

\item{label}{optional, Add an extra column with a label}

\item{origin}{optional, a table with sequence id numbers and their database origins}
}
\value{

}
\description{
summarise_fasta
}
