% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{Summarise phyloseq table}
\usage{
summarize_taxa(physeq, Rank, GroupBy = NULL)
}
\arguments{
\item{GroupBy}{}
}
\value{

}
\description{
Summarise phyloseq table
}
