% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{tax2phylo}
\alias{tax2phylo}
\title{tax2phylo}
\usage{
tax2phylo(
  x,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  depth = "all",
  summarise = FALSE,
  append_sum = TRUE,
  hex = FALSE,
  resolve_poly = FALSE
)
}
\arguments{
\item{x}{a DNAbin object or an object coercible to DNAbin}

\item{ranks}{The taxonomic ranks currently assigned to the names}

\item{depth}{The depth within the ranks to return. Default all to return lowest input rank}

\item{summarise}{Summarise the number of sequences at depth.}

\item{append_sum}{Append the summary number to the output trees tip labels when summarise is TRUE}

\item{hex}{Convert the accessions to hexadecimal format to avoid newick incompatible characters. Can be reversed with hex2acc}

\item{resolve_poly}{Randomly resolve polytomies in tree using \code{ape::multi2di}. use 'all' to resolve all polytomies, or 'upper' to only resolve upper branches, useful for making constraints trees}
}
\description{
A newer, faster tax2tree that grafts tips. Only returns a phylo objec
}
