% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{train_idtaxa}
\alias{train_idtaxa}
\title{Train IDTAXA}
\usage{
train_idtaxa(
  x,
  maxGroupSize = 10,
  maxIterations = 3,
  allowGroupRemoval = TRUE,
  orient = FALSE,
  get_lineage = FALSE,
  db = NULL,
  quiet = FALSE
)
}
\arguments{
\item{x}{A DNAbin}

\item{maxGroupSize}{The}

\item{maxIterations}{The number of i}

\item{orient}{Training sequences must all be in the same orientation. Set this to TRUE to reorient the sequences if you are unsure.}

\item{get_lineage}{Get full taxonomic lineage using reformat_hierarchy if not already present}

\item{db}{A database file generated by \code{get_ncbi_lineage} or \code{get_ott_lineage}. Required if get_lineage is TRUE}

\item{quiet}{Whether progress should be printed to console}
}
\value{

}
\description{
Train IDTAXA
}
