% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAbin_utilities.R
\name{write_fasta}
\alias{write_fasta}
\title{write fasta}
\usage{
write_fasta(x, file = "", compress = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{a list of sequences in DNAbin or AAbin format, or a vector of sequences as concatenated upper-case character strings.}

\item{file}{character string giving a valid file path to output the text to. If file = "" (default setting) the text file is written to the console.}

\item{compress}{logical indicating whether the output file should be gzipped.}

\item{quiet}{Whether progress should be printed to consoe}
}
\value{

}
\description{
write fasta
}
