% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbmap_install}
\alias{bbmap_install}
\title{Install BBtools}
\usage{
bbmap_install(url, dest.dir = "bin", force = FALSE)
}
\arguments{
\item{url}{(Optional) Default will search for the latest version
URL to retrieve bbmap from.}

\item{dest.dir}{(Optional)  Default "bin"
Directory to install bbmap within.}
}
\value{

}
\description{
Install BBtools
}
