% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbtools_split}
\alias{bbtools_split}
\title{Split interleaved reads}
\usage{
bbtools_split(install = NULL, files, overwrite = FALSE)
}
\arguments{
\item{install}{(Required) Install location for bbmap}

\item{files}{(Required) Vector of locations of interleaved read files to split}

\item{overwrite}{(Optional) Default TRUE
Option to overwrite existing output files.}
}
\description{
Split interleaved reads
}
