% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{blast_install}
\alias{blast_install}
\title{Install BLAST}
\usage{
blast_install(url, dest.dir = "bin")
}
\arguments{
\item{url}{(Optional) Default will search for the latest version
URL to retrieve BLAST version from.}

\item{dest.dir}{(Optional)  Default "bin"
Directory to install BLAST within.}
}
\value{

}
\description{
Install BLAST
}
