% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{blast_top_hit}
\alias{blast_top_hit}
\title{BLAST Top Hit}
\usage{
blast_top_hit(
  query,
  db,
  type = "blastn",
  threshold = 90,
  taxranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  delim = ";",
  args = "-max_target_seqs 5",
  quiet = FALSE
)
}
\arguments{
\item{query}{(Required) Query sequence. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.}

\item{db}{(Required) Reference sequences to conduct search against. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.
If DNAbin, DNAStringSet or character string is provided, a temporary fasta file is used to construct BLAST database}

\item{type}{(Required) type of search to conduct, default 'blastn'}

\item{threshold}{(Required) Minimum identity threshold to accept}

\item{taxranks}{(Required) The taxonomic ranks contained in the fasta headers}

\item{delim}{(Required) The delimiter between taxonomic ranks in fasta headers}

\item{args}{(Optional) Extra arguments passed to BLAST}

\item{quiet}{(Optional) Whether progress should be printed to console, default is FALSE}
}
\value{

}
\description{
Conduct BLAST search and return top hit
}
