% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{blast_top_hit}
\alias{blast_top_hit}
\title{BLAST Top Hit}
\usage{
blast_top_hit(
  query,
  db,
  type = "blastn",
  identity = 95,
  coverage = 95,
  evalue = 1e+06,
  max_target_seqs = 5,
  max_hsp = 5,
  ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  delim = ";",
  tie = "first",
  args = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{query}{(Required) Query sequence. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.}

\item{db}{(Required) Reference sequences to conduct search against. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.
If DNAbin, DNAStringSet or character string is provided, a temporary fasta file is used to construct BLAST database}

\item{type}{(Required) type of search to conduct, default 'blastn'}

\item{identity}{(Required) Minimum percent identity cutoff.}

\item{coverage}{(Required) Minimum percent query coverage cutoff.}

\item{evalue}{(Required) Minimum expect value (E) for saving hits}

\item{max_target_seqs}{(Required) Number of aligned sequences to keep. Even if you are only looking for 1 top hit keep this higher for calculations to perform properly.}

\item{max_hsp}{(Required) Maximum number of HSPs (alignments) to keep for any single query-subject pair.}

\item{ranks}{(Required) The taxonomic ranks contained in the fasta headers}

\item{delim}{(Required) The delimiter between taxonomic ranks in fasta headers}

\item{tie}{How to handle ties in top hit results. Options are to break ties by selecting the first hit (Default), or return all tied hits.}

\item{args}{(Optional) Extra arguments passed to BLAST}

\item{quiet}{(Optional) Whether progress should be printed to console, default is FALSE}
}
\value{

}
\description{
Conduct BLAST search and return top hit
}
