% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_taxa.R
\name{boldSearch}
\alias{boldSearch}
\title{boldSearch}
\usage{
boldSearch(x, marker = "COI-5P", quiet = FALSE, output = "h",
  file = NULL, compress = FALSE, dir = NULL)
}
\arguments{
\item{dir}{}
}
\description{
boldSearch
}
