% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{boldSearch}
\alias{boldSearch}
\title{boldSearch}
\usage{
boldSearch(
  x,
  marker = "COI-5P",
  quiet = FALSE,
  output = "h",
  out.file = NULL,
  compress = FALSE,
  force = FALSE,
  out.dir = NULL,
  db = NULL
)
}
\arguments{
\item{x}{A taxon name or vector of taxa to download sequences for}

\item{marker}{the barcode marker used as a search term for the database}

\item{output}{the output format for the taxonomy in fasta headers.
Options include "h" for full heirarchial taxonomy (SeqID;Domain;Phylum;Class;Order;Family;Genus;Species),
"binom" for just genus species binomials (SeqID;Genus Species),
"bold" for BOLD taxonomic ID only (SeqID;BoldTaxID),
"gb" for genbank taxonomic ID (SeqID;GBTaxID),
or "gb-binom" which outputs Genus species binomials, as well as genbank taxonomic ID's, and translates all BOLD taxonomic ID's to genbank taxonomic ID's in the process}

\item{out.file}{The file to write to, if empty it defaults to the search term}

\item{compress}{Option to compress output fasta files using gzip}

\item{force}{Option ot overwright files if they already exist}

\item{out.dir}{Output directory to write fasta files to}
}
\value{

}
\description{
boldSearch
}
