% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{boldUpdate}
\alias{boldUpdate}
\title{boldUpdate}
\usage{
boldUpdate(
  x,
  fasta,
  marker = "COI-5P",
  quiet = FALSE,
  output = "h",
  suffix = "updates",
  compress = FALSE,
  force = FALSE,
  out.dir = NULL,
  db = NULL
)
}
\arguments{
\item{x}{A taxon name or vector of taxa to download sequences for}

\item{fasta}{A fasta file or list of fasta files to check for existing sequence accessions}

\item{marker}{the barcode marker used as a search term for the database}

\item{quiet}{(Optional) Print text output}

\item{output}{the output format for the taxonomy in fasta headers.
Options include "h" for full heirarchial taxonomy (SeqID;Domain;Phylum;Class;Order;Family;Genus;Species),
"binom" for just genus species binomials (SeqID;Genus Species),
"bold" for BOLD taxonomic ID only (SeqID;BoldTaxID),
"gb" for genbank taxonomic ID (SeqID;GBTaxID),
"gb-binom" which outputs Genus species binomials, as well as genbank taxonomic ID's, and translates all BOLD taxonomic ID's to genbank taxonomic ID's in the process
or "standard" which outputs the default format for each database. For bold this is \verb{processid|species name|markercode|genbankid}}

\item{suffix}{The suffix to add to newly downloaded files. Defaults to 'updates'}

\item{compress}{Option to compress output fasta files using gzip}

\item{force}{Option ot overwright files if they already exist}

\item{out.dir}{Output directory to write fasta files to}

\item{db}{}

\item{out.file}{The file to write to, if empty it defaults to the search term}
}
\value{

}
\description{
boldUpdate
}
