% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{clean_seqs}
\alias{clean_seqs}
\title{Clean Sequences with PHMM}
\usage{
clean_seqs(
  x,
  model,
  minscore = 100,
  shave = TRUE,
  maxNs = 0,
  cores = 1,
  quiet = FALSE,
  progress = FALSE,
  ...
)
}
\arguments{
\item{x}{A DNAbin or DNAStringset object}

\item{model}{A profile hidden Markov model (a "PHMM" object) generated by the aphid R package to align the sequences to.}

\item{minscore}{The minimum specificity (log-odds score for the optimal alignment) between the query sequence and the PHMM model for the sequence to be retained in the output object.}

\item{shave}{Whether bases that are outside (to the left or right) of the PHMM object should be shaved from the outputs.}

\item{maxNs}{The max number of ambiguous N bases to retain}

\item{cores}{The number of CPUs to parallelize the operation over}

\item{quiet}{Whether progress should be printed to the console.}
}
\value{

}
\description{
Clean Sequences with PHMM
}
