% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{codon_entropy}
\alias{codon_entropy}
\title{Codon entropy}
\usage{
codon_entropy(
  x,
  genetic.code = "SGC4",
  forward = TRUE,
  reverse = FALSE,
  codon.filter = TRUE,
  method = "ML"
)
}
\arguments{
\item{codon.filter}{}
}
\value{

}
\description{
Codon entropy
}
