% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{codon_filter}
\alias{codon_filter}
\title{Filter sequences containing stop codons}
\usage{
codon_filter(x, genetic.code = "SGC4", forward = TRUE,
  reverse = FALSE)
}
\arguments{
\item{reverse}{}
}
\description{
Filter sequences containing stop codons
}
