% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{codon_filter}
\alias{codon_filter}
\title{Filter sequences containing stop codons}
\usage{
codon_filter(x, genetic_code = NULL, tryrc = TRUE, resolve_draws = "majority")
}
\arguments{
\item{x}{Sequences in DNAStringset or DNAbin format}

\item{genetic_code}{A genetic code for the Amino acid translation. set to 'SGC4' for Invertebrate mitochondrial or see all known codes at Biostrings::GENETIC_CODE_TABLE}

\item{tryrc}{Whether the reverse complemement should be evaluated if no frame without stop codons was found in the forward orientation.}

\item{resolve_draws}{How draws should be resolved when multiple possible frames produce sequences with no stop codons.
Options are "remove" to completely remove the sequence, or "majority" to pick the most common frame from the entire alignment.}
}
\description{
Filter sequences containing stop codons
}
