% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{codon_filter}
\alias{codon_filter}
\title{Filter sequences containing stop codons}
\usage{
codon_filter(x, genetic.code = "SGC4", forward = TRUE,
  reverse = FALSE, remove.ambiguities = TRUE)
}
\arguments{
\item{x}{Sequences in DNAStringset or DNAbin format}

\item{genetic.code}{A genetic code for the Amino acid translation. See all known codes at GENETIC_CODE_TABLE
Default is the invertebrate mitochondrial code 'SGC4'}

\item{forward}{Whether the forward complement should be used}

\item{reverse}{Whether the reverse complement should be used}

\item{remove.ambiguities}{Whether ambiguous bases (non ATGC) bases should be removed}
}
\description{
Filter sequences containing stop codons
}
