% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{gbSearch}
\alias{gbSearch}
\title{Genbank search function}
\usage{
gbSearch(
  x,
  database = "nuccore",
  marker = c("COI", "CO1", "COX1"),
  quiet = FALSE,
  output = "h",
  minlength = 1,
  maxlength = 2000,
  subsample = NULL,
  chunksize = NULL,
  out.dir = NULL,
  compress = FALSE,
  force = FALSE
)
}
\arguments{
\item{x}{A taxon name or vector of taxa to download sequences for}

\item{marker}{The barcode marker used as a search term for the database. If this is set to "mitochondria" it will download full mitochondrial genomes.}

\item{quiet}{(Optional) Print text output}

\item{output}{The output format for the taxonomy in fasta headers.
Options include "h" for full heirarchial taxonomy (SeqID;Domain;Phylum;Class;Order;Family;Genus;Species),
"binom" for just genus species binomials (SeqID;Genus Species),
"bold" for BOLD taxonomic ID only (SeqID;BoldTaxID),
"gb" for genbank taxonomic ID (SeqID;GBTaxID),
or "gb-binom" which outputs Genus species binomials, as well as genbank taxonomic ID's, and translates all BOLD taxonomic ID's to genbank taxonomic ID's in the process}

\item{minlength}{The minimum length of sequences to download}

\item{maxlength}{The maximum length of sequences to download}

\item{out.dir}{Output directory to write fasta files to}

\item{compress}{Option to compress output fasta files using gzip}

\item{force}{Option ot overwright files if they already exist}
}
\value{

}
\description{
Genbank search function
}
