% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{get_mixed_clusters}
\alias{get_mixed_clusters}
\title{Get mixed clusters}
\usage{
get_mixed_clusters(
  x,
  db,
  rank = "order",
  threshold = 0.97,
  confidence = 0.8,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{A DNAbin list object whose names include NCBItaxonomic identification numbers.}

\item{db}{A NCBI taxonomy database from get_ncbi_lineage. If missing will download new one}

\item{rank}{The taxonomic rank to check clusters at, accepts a character such as "order", or vector of characters such as c("species", "genus").
If "all", the clusters will be checked at all taxonomic ranks available.}

\item{threshold}{numeric between 0 and 1 giving the OTU identity cutoff for clustering. Defaults to 0.97.}

\item{confidence}{The minimum confidence value for a sequence to be purged. For example, if confidence = 0.8 (the default value)
a sequence will only be purged if its taxonomy differs from at least four other independent sequences in its cluster.}

\item{quiet}{logical indicating whether progress should be printed to the console.}

\item{...}{further arguments to pass to kmer::otu (not including nstart).}
}
\value{

}
\description{
Get mixed clusters
}
