% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_ncbi_lineage}
\alias{get_ncbi_lineage}
\title{Download NCBI taxdump}
\usage{
get_ncbi_lineage(db = "NCBI", synonyms = TRUE, force = FALSE)
}
\arguments{
\item{synonyms}{}
}
\value{

}
\description{
Download NCBI taxdump
}
