% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_ncbi_synonyms}
\alias{get_ncbi_synonyms}
\title{Get NCBI synonyms}
\usage{
get_ncbi_synonyms(dir = NULL, recurse = TRUE, quiet = FALSE)
}
\arguments{
\item{dir}{A directory containing the NCBI taxonomy that was downloaded using get_ncbi_taxonomy()}

\item{recurse}{Whether to recurse when searching for dir if dir is NULL.
If TRUE recurse fully, if a positive number the number of levels to recurse.}

\item{quiet}{Whether progress should be printed to the console.}
}
\value{

}
\description{
Get NCBI synonyms
}
