% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{get_ott_lineage}
\alias{get_ott_lineage}
\title{Recursively get lineage from OTT taxid
This function derives the full lineage of a taxon ID number from a given taxonomy database}
\usage{
get_ott_lineage(
  x,
  db,
  output = "tax_name",
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  cores = 1
)
}
\arguments{
\item{x}{A DNAbin}

\item{ranks}{the taxonomic ranks to filter to. Default is "kingdom", "phylum", "class", "order", "family", "genus", "species"
To get strain level ranks, add "terminal" to ranks}

\item{cores}{integer giving the number of CPUs to parallelize the operation over (Defaults to 1).
This argument may alternatively be a 'cluster' object, in which case it is the user's responsibility to close the socket connection at the conclusion of the operation,
for example by running parallel::stopCluster(cores).
The string 'autodetect' is also accepted, in which case the maximum number of cores to use is one less than the total number of cores available.
Note that in this case there may be a tradeoff in terms of speed depending on the number and size of sequences to be processed, due to the extra time required to initialize the cluster.}

\item{dir}{a directory containing the open tree of life taxonomy files obtained from the  \code{download_ott_taxonomy} function}
}
\value{

}
\description{
Recursively get lineage from OTT taxid
This function derives the full lineage of a taxon ID number from a given taxonomy database
}
