% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{get_ott_lineage}
\alias{get_ott_lineage}
\title{Recursively get lineage from OTT taxid
This function derives the full lineage of a taxon ID number from a given taxonomy database}
\usage{
get_ott_lineage(
  x,
  db,
  output = "tax_name",
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  multithread = FALSE
)
}
\arguments{
\item{x}{A DNAbin, DNAStringSet, taxonomy headers formnatted Acc|taxid;taxonomy, or vector of taxids}

\item{db}{an OTT taxonomic database}

\item{output}{Options are "tax_name" to return taxonomic names for each rank in the lineage or "tax_id" to return taxonomic ID's.}

\item{ranks}{the taxonomic ranks to filter to. Default is "kingdom", "phylum", "class", "order", "family", "genus", "species"
To get strain level ranks, add "terminal" to ranks}

\item{multithread}{Whether multithreading should be used, if TRUE the number of cores will be automatically detected (Maximum available cores - 1), or provide a numeric vector to manually set the number of cores to use. Default is FALSE (single thread)
This argument may alternatively be a 'cluster' object, in which case it is the user's responsibility to close the socket connection at the conclusion of the operation,
for example by running parallel::stopCluster(cores).}
}
\description{
Recursively get lineage from OTT taxid
This function derives the full lineage of a taxon ID number from a given taxonomy database
}
