% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_ranked_lineage}
\alias{get_ranked_lineage}
\title{(Deprecated) Get ranked Lineage}
\usage{
get_ranked_lineage(db = "NCBI", synonyms = TRUE, force = FALSE)
}
\arguments{
\item{force}{}
}
\value{

}
\description{
(Deprecated) Get ranked Lineage
}
