% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{get_reading_frame}
\alias{get_reading_frame}
\title{Get Reading frame of sequences}
\usage{
get_reading_frame(x, genetic.code = "SGC4", forward = TRUE,
  reverse = FALSE)
}
\arguments{
\item{x}{Sequences in DNAStringset or DNAbin format}

\item{genetic.code}{A genetic code for the Amino acid translation. See all known codes at GENETIC_CODE_TABLE
Default is the invertebrate mitochondrial code 'SGC4'}

\item{reverse}{}
}
\description{
Get Reading frame of sequences
}
