% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{gid_to_acc}
\alias{gid_to_acc}
\title{Convert NCBI gene ids to accession numbers}
\usage{
gid_to_acc(
  ids,
  database = "nuccore",
  chunksize = 300,
  multithread = TRUE,
  progress = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{ids}{A character or numeric vector of NCBI gids}

\item{database}{The origin database of the gids, default 'nuccore'}

\item{chunksize}{The size of the chunked searches to conduct.
Warning, chunk sizes over 300 can be too big for the NCBI servers.}

\item{multithread}{Whether multithreading should be used}

\item{quiet}{(Optional) Print text output}
}
\value{

}
\description{
Convert NCBI gene ids to accession numbers
}
