% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{lca_probs}
\alias{lca_probs}
\title{Probabilitys of sharing a rank as a function of sequence identity}
\usage{
lca_probs(
  x,
  sim = seq(0.9, 1, 0.01),
  k = 5,
  nstart = 20,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  delim = ";"
)
}
\arguments{
\item{x}{a DNAbin object or an object coercible to DNAbin}

\item{sim}{The sequence similarities to cluster at}

\item{k}{integer giving the k-mer size used to generate the input matrix for k-means clustering.}

\item{nstart}{value passed to  nstart passed to kmeans. Higher increases computation time but can improve clustering accuracy considerably.}

\item{ranks}{The taxonomic ranks currently assigned to the names}

\item{delim}{The delimiter used between ranks}
}
\value{

}
\description{
Probabilitys of sharing a rank as a function of sequence identity
}
