% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{make_blast_db}
\alias{make_blast_db}
\title{Make blast Database}
\usage{
make_blast_db(file, dbtype = "nucl", args = NULL, quiet = FALSE)
}
\arguments{
\item{file}{(Required) A fasta file to create a database from.}

\item{dbtype}{(Optional) Molecule type of database, accepts "nucl" for nucleotide or "prot" for protein.}

\item{args}{(Optional) Extra arguments passed to BLAST}

\item{quiet}{(Optional) Whether progress should be printed to console, default is FALSE}
}
\value{

}
\description{
Make blast Database
}
