% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Make blast Database}
\usage{
makeblastdb(file, dbtype = "nucl", args = NULL)
}
\arguments{
\item{file}{(Required) A fasta file to create a database from.}

\item{dbtype}{(Optional) Type of database, accepts "nucl" for nucleotide
default is "nucl" for nucleotide}

\item{args}{(Optional) Extra arguments passed to BLAST}
}
\description{
Make blast Database
}
