% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Make blast Database}
\usage{
makeblastdb(file, dbtype = "nucl", args = NULL)
}
\arguments{
\item{file}{A fasta file to create a database from}

\item{dbtype}{Type of database, default is nucleotide}

\item{args}{Extra arguments passed to BLAST}
}
\description{
Make blast Database
}
