% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_utilities.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Make blast Database}
\usage{
makeblastdb(file, dbtype = "nucl", args = NULL, quiet = FALSE)
}
\arguments{
\item{file}{(Required) A fasta file to create a database from.}

\item{dbtype}{(Optional) Molecule type of database, accepts "nucl" for nucleotide or "prot" for protein.}

\item{args}{(Optional) Extra arguments passed to BLAST}
}
\value{

}
\description{
Make blast Database
}
