% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences.R
\name{map_to_model}
\alias{map_to_model}
\title{Map sequences to PHMM}
\usage{
map_to_model(
  x,
  model,
  minscore = 100,
  shave = TRUE,
  check_indels = TRUE,
  extra = NULL,
  maxNs = Inf,
  multithread = FALSE,
  quiet = FALSE,
  progress = FALSE,
  ...
)
}
\arguments{
\item{x}{A DNAbin or DNAStringset object}

\item{model}{A profile hidden Markov model (a "PHMM" object) generated by \code{aphid::derivePHMM} to align the sequences to.}

\item{minscore}{The minimum specificity (log-odds score for the optimal alignment) between the query sequence and the PHMM model for the sequence to be retained in the output object.}

\item{shave}{Whether bases that are outside (to the left or right) of the PHMM object should be shaved from sequences in the output.}

\item{check_indels}{Whether sequences with indels not in multiples of 3 should be removed from output. Useful for coding regions such as COI.}

\item{extra}{How to handle extra length sequences when insertions that were not part of the model appear. 'drop' will truncate all sequences to the shortest alignment, 'fill' will fill with gaps}

\item{maxNs}{The max number of ambiguous N bases to retain}

\item{multithread}{Whether multithreading should be used, if TRUE the number of cores will be automatically detected, or provided a numeric vector to manually set the number of cores to use}

\item{quiet}{Whether progress should be printed to the console.}

\item{...}{Any further arguments to be passed to \code{aphid::Viterbi}}
}
\value{

}
\description{
Map sequences to PHMM
}
