% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott_taxonomy.R
\name{map_to_ott}
\alias{map_to_ott}
\title{Map taxa to open tree of life}
\usage{
map_to_ott(
  x,
  dir = NULL,
  from = "ncbi",
  resolve_synonyms = TRUE,
  filter_bads = TRUE,
  remove_na = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{x}{a DNAbin}

\item{dir}{a directory containing the open tree of life taxonomy files obtained from the  \code{download_ott_taxonomy} function}

\item{from}{the existing taxonomic ID format}

\item{resolve_synonyms}{Whether to resolve synonyms}

\item{filter_bads}{Whether to filter 'bad' entries. These include
incertae_sedis
major_rank_conflict
unplaced
environmental
inconsistent
extinct
hidden
hybrid
not_otu
viral
barren
See: https://github.com/OpenTreeOfLife/reference-taxonomy/blob/master/doc/taxon-flags.md for more info}

\item{remove_na}{Whether taxa that could not be mapped to the open tree of life should be removed}

\item{quiet}{Whether progress should be printed to console}
}
\value{

}
\description{
Map taxa to open tree of life
}
