% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pad_alignment}
\alias{pad_alignment}
\title{Pad alignment}
\usage{
pad_alignment(
  x,
  model,
  pad = "-",
  tryrc = FALSE,
  quiet = FALSE,
  check_indels = TRUE,
  minscore = 10,
  ...
)
}
\arguments{
\item{x}{A DNAbin object or an object coercible to DNAbin}

\item{model}{A profile hidden Markov model (a "PHMM" object) generated by the aphid R package to align the sequences to.}

\item{pad}{The character used to pad the gaps}

\item{tryrc}{Whether the reverse complement should also be aligned}

\item{quiet}{Whether progress should be printed to the console.}

\item{check_indels}{Check that indels are multiples of 3, recommended for coding sequences such as COI}

\item{minscore}{Minimum score for the viterbi alignment}

\item{...}{aditional arguments to be passed to "Viterbi"}
}
\value{

}
\description{
Aligns a DNABin to a reference PHMM model, and pads any gaps between the query and reference
}
