% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{prune_groups}
\alias{prune_groups}
\title{Prune group sizes}
\usage{
prune_groups(x, maxGroupSize = 5, dedup = TRUE, discardby = "random",
  quiet = FALSE)
}
\arguments{
\item{quiet}{}
}
\description{
Prune group sizes
}
