% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ps_to_fasta}
\alias{ps_to_fasta}
\title{outputs a FASTA file from a phyloseq object}
\usage{
ps_to_fasta(ps = ps, out.file = NULL, rank = NULL, width = 1000, ...)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains \code{\link[phyloseq]{refseq}}.
If there the \code{refseq} slot is not filled, this function will try pull the
sequences from \code{\link[phyloseq]{get_taxa}}}

\item{rank}{(optional) A taxonomic rank from the \code{\link[phyloseq]{tax_table}} which will be used to name the sequences.
If no rank is supplied, samples will be named \code{ASV_#}}

\item{width}{(Default 1000) The number of characters in each fasta line before wrapping occurs}

\item{...}{(Optional) Any further paramaters to be passed to \code{writeXStringSet}}

\item{file}{(optional) A file name that ends in ".fasta" or ".fa".
If a file name is not supplied, the file will be named after the phyloseq object.}
}
\value{
This function saves a FASTA-formatted text file from the input \code{phyloseq} object.
}
\description{
This function outputs a FASTA-formatted text file from a \code{phyloseq} object.
This code was modified from \code{reltools} package https://github.com/DanielSprockett/reltools by Daniel Sprocket
}
\examples{
save_fasta(ps)
save_fasta(ps = ps, file = "sequences.fasta", rank = "Genus")
}
