% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_dada2_gen}
\alias{reformat_dada2_gen}
\title{Reformat DADA2 Genus}
\usage{
reformat_dada2_gen(x, db = NULL, quiet = FALSE, ranks = NULL, force = FALSE)
}
\arguments{
\item{ranks}{}
}
\value{

}
\description{
Reformat DADA2 Genus
}
