% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_db.R
\name{reformat_dada2_gen}
\alias{reformat_dada2_gen}
\title{Reformat DADA2 Genus}
\usage{
reformat_dada2_gen(x, db = NULL, quiet = TRUE, ranks = NULL)
}
\arguments{
\item{ranks}{}
}
\description{
Reformat DADA2 Genus
}
