% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_dada2_spp}
\alias{reformat_dada2_spp}
\title{Reformat DADA2 Species}
\usage{
reformat_dada2_spp(x, quiet = FALSE)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Reformat DADA2 Species
}
