% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_db.R
\name{reformat_heirarchy}
\alias{reformat_heirarchy}
\title{Reformat heirarchy}
\usage{
reformat_heirarchy(x, db = NULL, quiet = TRUE, ranks = NULL,
  sppsep = "_")
}
\arguments{
\item{ranks}{}
}
\description{
Reformat heirarchy
}
