% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_hierarchy}
\alias{reformat_hierarchy}
\title{Reformat hierarchy}
\usage{
reformat_hierarchy(
  x,
  db = NULL,
  quiet = FALSE,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  cores = 1
)
}
\arguments{
\item{x}{A DNAbin with names formatted Accession|taxid;taxonomy}

\item{quiet}{Whether progress should be printed to console}

\item{ranks}{The taxonomic ranks to return in the output names}

\item{cores}{The number of  cores to use for get_ott_lineage}
}
\value{

}
\description{
Reformat hierarchy
}
