% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{reformat_hierarchy}
\alias{reformat_hierarchy}
\title{Reformat hierarchy}
\usage{
reformat_hierarchy(
  x,
  db = NULL,
  quiet = FALSE,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  multithread = FALSE
)
}
\arguments{
\item{x}{A DNAbin with names formatted Accession|taxid;taxonomy}

\item{db}{A database file generated by \code{get_ncbi_taxonomy} or \code{get_ott_lineage}. Required if get_lineage is TRUE}

\item{quiet}{Whether progress should be printed to console}

\item{ranks}{The taxonomic ranks to return in the output names}

\item{multithread}{Whether multithreading should be used, if TRUE the number of cores will be automatically detected (Maximum available cores - 1), or provide a numeric vector to manually set the number of cores to use. Default is FALSE (single thread)
This argument may alternatively be a 'cluster' object, in which case it is the user's responsibility to close the socket connection at the conclusion of the operation,
for example by running parallel::stopCluster(cores).}
}
\value{

}
\description{
Reformat hierarchy
}
