% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa.R
\name{resolve_taxonomy}
\alias{resolve_taxonomy}
\title{Resolve taxonomic synonyms}
\usage{
resolve_taxonomy(
  x,
  subspecies = FALSE,
  quiet = TRUE,
  missing = "ignore",
  higherrank = FALSE,
  fuzzy = TRUE
)
}
\arguments{
\item{x}{A DNAbin or DNAStringset object}

\item{subspecies}{Whether subspecies should be included}

\item{quiet}{Whether progress should be printed to the console.}

\item{missing}{How to handle cases where the updated taxonomic name does not have a taxonomic ID in the NCBI database.
Options include "ignore", which will update name but keep old taxid, "keepold" which will keep the old taxname and taxid,
and "remove" which will remove all synonyms that dont have a taxid in the NCBI database.}

\item{higherrank}{Whether taxa that were not found should be mapped to a higher rank}

\item{fuzzy}{Whether fuzzy matching of taxa names should be used}
}
\value{
This returns a DNAbin with renamed taxa
}
\description{
This function takes a DNAbin object, or a list of species and uses the Global Biodiversity Information Facility (GBIF) to resolve taxonomic synonyms
}
