% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{split_bold_query}
\alias{split_bold_query}
\title{Split bold query
This function recursively splits a bold taxonomic query until the amount of records returned is under chunk_size}
\usage{
split_bold_query(x, chunk_size = 1e+05, quiet = FALSE)
}
\arguments{
\item{x}{The input taxonomic query}

\item{chunk_size}{The maximum amount of records to return per query}

\item{quiet}{Whether progress should be printed to the console.}
}
\value{

}
\description{
Split bold query
This function recursively splits a bold taxonomic query until the amount of records returned is under chunk_size
}
