% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{tax2tree}
\alias{tax2tree}
\title{tax2tree}
\usage{
tax2tree(
  x,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  depth = NULL,
  summarise = FALSE,
  output = "phylo",
  replace_bads = FALSE
)
}
\arguments{
\item{x}{a DNAbin object or an object coercible to DNAbin}

\item{ranks}{The taxonomic ranks currently assigned to the names}

\item{depth}{The depth within the tree to return. Default NULL to return lowest input rank}

\item{summarise}{Select a taxonomic rank to summarise below. Default is FALSE to return a tree of the same size as input}

\item{output}{The output to return, options are:
"phylo" to return an ape phylo object
"data.tree" to return a data.tree node object
"treedf" to return a simplified tree with taxon summaries in a data frame
"newick" to return a newick file for visualisation in other software}

\item{replace_bads}{An option to automatically replace invalid newick characters with '-'}
}
\value{
a phylo, data.tree, newick, or treedf object
}
\description{
tax2tree
}
