% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{train_idtaxa}
\alias{train_idtaxa}
\title{Train IDTAXA}
\usage{
train_idtaxa(x, maxGroupSize = 10, maxIterations = 3,
  allowGroupRemoval = TRUE, db = NULL, quiet = FALSE,
  force = FALSE)
}
\arguments{
\item{force}{}
}
\description{
Train IDTAXA
}
