% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_utilities.R
\name{train_idtaxa}
\alias{train_idtaxa}
\title{Train IDTAXA}
\usage{
train_idtaxa(
  x,
  maxGroupSize = 10,
  maxIterations = 3,
  allowGroupRemoval = TRUE,
  db = NULL,
  quiet = FALSE,
  force = FALSE,
  get_lineage = FALSE
)
}
\arguments{
\item{get_lineage}{Reformat the sequences heirarchially before training}
}
\value{

}
\description{
Train IDTAXA
}
